// interfaces/comercial/movimientos/amortizarCompraInterface.ts
export interface IAmortizarCompra {
    girado: string
    tipo_documento_pago_id: string
    fecha_pago: string
    cuenta: string
    monto_a_pagar: string
    nro_de_documento_pago: string
    tipo_de_pago_compra: string
    moneda_id: string
    glosa: string
    compra_id: string
    proveedor: string
    tipo_documento_compra_id: string
    nro_documento_compra: string
    deuda: string
}